pyproject.toml
[project]
name = "collective.person"
dynamic = ["version"]
description = "A Person content type for Plone."
readme = "README.md"
license = {text = "GNU General Public License v2 (GPLv2)"}
requires-python = ">=3.10"
authors = [
    { name = "Érico Andrei", email = "ericof@plone.org" },
]
keywords = [
    "Plone",
    "Python",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Framework :: Plone",
    "Framework :: Plone :: 6.0",
    "Framework :: Plone :: 6.1",
    "Framework :: Plone :: Addon",
    "Framework :: Zope",
    "Framework :: Zope :: 5",
    "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "Products.CMFPlone>=6.0.0",
    "plone.api",
    "plone.restapi",
    "collective.contact_behaviors>=1.0.0b3",
    "plone.volto",
    "plone.exportimport",
]

[project.optional-dependencies]
test = [
    "horse-with-no-namespace>=20250705.0",
    "plone.app.testing",
    "plone.restapi[test]",
    "pytest-cov",
    "pytest-plone>=1.0.0a1",
    "zest-releaser[recommended]>=9.5.0",
    "zestreleaser-towncrier>=1.3.0",
]

[project.urls]
Changelog = "https://github.com/collective/collective.person/blob/main/CHANGELOG.md"
Homepage = "https://github.com/collective/collective.person"
Issues = "https://github.com/collective/collective.person/issues"
Repository = "https://github.com/collective/collective.person/"


[project.entry-points."z3c.autoinclude.plugin"]
target = "plone"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/collective/person/__init__.py"

[tool.hatch.build]
strict-naming = true

[tool.hatch.build.targets.sdist]
include = [
    "/*.md",
    "/*.yaml",
    "/*.yml",
    "/docs",
    "/src",
    "/tests",
]
exclude = [
  "/.github",
]

[tool.hatch.build.targets.wheel]
packages = ["src/collective", ]

[tool.uv]
managed = false

[tool.towncrier]
package = "collective.person"
package_dir = "src"
directory = "news/"
filename = "CHANGELOG.md"
start_string = "<!-- towncrier release notes start -->\n"
issue_format = "[#{issue}](https://github.com/collective/collective.person/issues/{issue})"
title_format = "## {version} ({project_date})"
template = "news/.changelog_template.jinja"
underlines = ["", "", ""]

[[tool.towncrier.type]]
directory = "breaking"
name = "Breaking changes:"
showcontent = true

[[tool.towncrier.type]]
directory = "feature"
name = "New features:"
showcontent = true

[[tool.towncrier.type]]
directory = "bugfix"
name = "Bug fixes:"
showcontent = true

[[tool.towncrier.type]]
directory = "internal"
name = "Internal:"
showcontent = true

[[tool.towncrier.type]]
directory = "documentation"
name = "Documentation:"
showcontent = true

[[tool.towncrier.type]]
directory = "tests"
name = "Tests"
showcontent = true

[tool.black]
enable-unstable-feature = [
    "hug_parens_with_braces_and_square_brackets"
]

[tool.ruff]
target-version = "py310"
line-length = 88
fix = true
lint.select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
]
lint.ignore = [
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.isort]
case-sensitive = false
no-sections = true
force-single-line = true
from-first = true
lines-after-imports = 2
lines-between-types = 1
order-by-type = false

[tool.ruff.format]
preview = true

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E501", "RUF001", "S101"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source_pkgs = ["collective.person", "tests"]
branch = true
parallel = true
omit = [
  "src/collective/person/locales/*",
]

[tool.zest-releaser]
upload-pypi = false # Build and upload with uv
python-file-with-version = "src/collective/person/__init__.py"
